fix: move test to test_utils + tests convert_to_geojson_polytope
