+++ eodag/api/product/metadata_mapping.py
@@ -181,6 +181,7 @@ def format_metadata(search_param: str, *args: Any, **kwargs: Any) -> str:
         - ``to_datetime_dict``: convert a datetime string to a dictionary where values are either a string or a list
         - ``to_ewkt``: convert to EWKT (Extended Well-Known text)
         - ``to_geojson``: convert to a GeoJSON (via __geo_interface__ if exists)
+        - ``to_geojson_polytope``: convert shapely Point/LineString/Polygon to ECMWF polytope feature dicts
         - ``to_iso_date``: remove the time part of a iso datetime string
         - ``to_iso_utc_datetime_from_milliseconds``: convert a utc timestamp in given milliseconds to a utc iso datetime
         - ``to_iso_utc_datetime``: convert a UTC datetime string to ISO UTC datetime string
@@ -434,6 +435,7 @@ def format_metadata(search_param: str, *args: Any, **kwargs: Any) -> str:
         def convert_to_geojson_polytope(
             value: BaseGeometry,
         ) -> Union[dict[Any, Any], str]:
+            """Convert a shapely Point/LineString/Polygon to ECMWF polytope feature dicts"""
             # ECMWF Polytope uses non-geojson structure for features
             if isinstance(value, Polygon):
                 return {
